requirements.txt
streamlit==1.28.1
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.2
openai-whisper==20240314
TTS==0.22.0
sounddevice==0.4.6
indic-transliteration==2.3.57
